@@ -6,27 +6,28 @@
   "repository": "silverwind/versions",
   "license": "BSD-2-Clause",
   "bin": "./versions",
+  "type": "module",
   "scripts": {
     "test": "make test"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "files": [
-    "versions"
+    "versions.cjs"
   ],
   "devDependencies": {
-    "@vercel/ncc": "0.27.0",
-    "eslint": "7.21.0",
-    "eslint-config-silverwind": "29.0.0",
+    "@babel/core": "7.13.16",
+    "@vercel/ncc": "0.28.3",
+    "eslint": "7.24.0",
+    "eslint-config-silverwind": "32.3.3",
     "execa": "5.0.0",
     "fast-glob": "3.2.5",
-    "jest": "26.6.3",
+    "jest": "27.0.0-next.8",
     "minimist": "1.2.5",
-    "updates": "11.4.3"
+    "updates": "12.0.2"
   },
   "keywords": [
-    "ver",
     "version",
     "version bump",
     "versions",
@@ -37,6 +38,8 @@
     "replace"
   ],
   "jest": {
+    "testTimeout": 60000,
+    "transform": {},
     "verbose": false
   }
 }
